refactor : 테스트 코드 클래스 추출 리팩토링
