+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,0 +1,3 @@
+package blackjack.domain
+
+data class Card(val rank: Ranks, val suits: Suits)
