+++ src/main/kotlin/blackjack/domain/Suits.kt
@@ -1,0 +1,10 @@
+package blackjack.domain
+
+enum class Suits(
+    val koreanName: String,
+) {
+    SPADE("스페이드"),
+    HEART("하트"),
+    DIAMOND("다이어"),
+    CLUB("클로버"),
+}
