+++ src/test/kotlin/blackjack/BlackJackTest.kt
@@ -1,10 +1,15 @@
 package blackjack
 
+import blackjack.domain.BlackJackGame
+import blackjack.domain.Card
+import blackjack.domain.GameCards
+import blackjack.domain.Player
+import blackjack.domain.Players
+import blackjack.domain.Ranks
+import blackjack.domain.Suits
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
-import java.util.LinkedList
-import java.util.Queue
 
 class BlackJackTest {
     @Test
@@ -97,234 +102,4 @@ class BlackJackTest {
             it.cardSize() shouldBe 2
         }
     }
-}
-
-class BlackJackGame(
-    private val players: Players,
-    private val gameCards: GameCards,
-) {
-    fun startGame() {
-        repeat(FIRST_ROUND_HAND_SIZE) {
-            giveOutCards()
-        }
-    }
-
-    private fun giveOutCards() {
-        players.forEach { players ->
-            players.receiveCard(gameCards.drawCard())
-        }
-    }
-
-    fun handlePlayerDraw(player: Player, gameCards: GameCards) {
-        val previousCardCount = player.cardSize()
-        InputView.enterIsContinueDrawCard(player)
-        if (player.isDrawContinue) {
-            player.receiveCard(gameCards.drawCard())
-            OutputView.printPlayerCard(player)
-            handlePlayerDraw(player, gameCards) // 재귀 호출
-        }
-
-        if (!player.isDrawContinue && player.cardSize() == previousCardCount) {
-            OutputView.printPlayerCard(player)
-        }
-    }
-
-    companion object {
-        private const val FIRST_ROUND_HAND_SIZE = 2
-    }
-}
-
-object InputView {
-    private const val CONTINUE_OR_STOP_MESSAGE = "는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)"
-
-    fun enterParticipatingPlayers(): Players {
-        println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)")
-        val playerNames = readln()
-        return Players.create(playerNames)
-    }
-
-    fun enterIsContinueDrawCard(player: Player) {
-        val userInput = getIsContinueDraw(player)
-        when (userInput.lowercase()) {
-            "n" -> player.stopCardDraw()
-            "y" -> player.continueCardDraw()
-            else -> println("잘못된 입력입니다. 다시 입력해주세요.")
-        }
-    }
-
-    fun getIsContinueDraw(player: Player): String {
-        println()
-        println(player.name + CONTINUE_OR_STOP_MESSAGE)
-        return readln()
-    }
-}
-
-object OutputView {
-    private const val PLAYER_NAME_DELIMITER = ", "
-    private const val BLANK_PREFIX_MESSAGE = ""
-    private const val POST_MESSAGE = "에게 2장의 나누었습니다."
-    private const val RESULT_EXPRESSION = " - 결과: "
-    private const val NAME_POSTFIX_EXPRESSION = "카드: "
-
-    fun printFirstAllPlayersCards(players: Players) {
-        println()
-        val result =
-            players.getPlayerNames().joinToString(
-                PLAYER_NAME_DELIMITER,
-                BLANK_PREFIX_MESSAGE,
-                POST_MESSAGE,
-            )
-        println(result)
-        players.forEach { player ->
-            printPlayerAllCards(player)
-            println()
-        }
-    }
-
-    fun printFinalResults(players: Players) {
-        println()
-        println()
-        println("게임 종료!")
-        players.forEach { player ->
-            printPlayerAllCards(player)
-            resultExpression(player)
-        }
-    }
-
-    fun printPlayerCard(player: Player) {
-        printPlayerAllCards(player)
-    }
-
-    private fun printPlayerAllCards(player: Player) {
-        print(player.name + NAME_POSTFIX_EXPRESSION)
-        print(player.findAllCardsNames().joinToString(PLAYER_NAME_DELIMITER))
-    }
-
-    private fun resultExpression(player: Player) {
-        print(RESULT_EXPRESSION)
-        println(player.calculateCardPoints())
-    }
-}
-
-data class Card(val rank: Ranks, val suits: Suits)
-
-enum class Ranks(
-    val keyword: String,
-    val points: List<Int>,
-) {
-    ACE("A", listOf(1, 11)),
-    TWO("2", listOf(2)),
-    THREE("3", listOf(3)),
-    FOUR("4", listOf(4)),
-    FIVE("5", listOf(5)),
-    SIX("6", listOf(6)),
-    SEVEN("7", listOf(7)),
-    EIGHT("8", listOf(8)),
-    NINE("9", listOf(9)),
-    TEN("10", listOf(10)),
-    JACK("J", listOf(10)),
-    QUEEN("Q", listOf(10)),
-    KING("K", listOf(10)),
-}
-
-enum class Suits(
-    val koreanName: String,
-) {
-    SPADE("스페이드"),
-    HEART("하트"),
-    DIAMOND("다이어"),
-    CLUB("클로버"),
-}
-
-class UserCards(private val cards: MutableList<Card>) : Collection<Card> by cards {
-    fun calculatePoints(): Int {
-        val basePoints = cards.sumOf { it.rank.points[0] }
-        val hasAce = cards.any { it.rank == Ranks.ACE }
-        if (hasAce && basePoints + 10 <= 21) {
-            return basePoints + 10
-        }
-        return basePoints
-    }
-
-    fun addCard(card: Card) {
-        cards.add(card)
-    }
-}
-
-class GameCards private constructor(private val deck: Queue<Card>) {
-    fun drawCard(): Card {
-        return deck.poll()
-    }
-
-    fun size(): Int {
-        return deck.size
-    }
-
-    companion object {
-        fun create(): GameCards {
-            val allCards =
-                Suits.entries.flatMap { suit ->
-                    Ranks.entries.map { rank -> Card(rank, suit) }
-                }
-            return GameCards(LinkedList(allCards.shuffled()))
-        }
-    }
-}
-
-class Player(val name: String, var isDrawContinue: Boolean = true) {
-    private var userCards = UserCards(mutableListOf())
-
-    fun receiveCard(card: Card) {
-        require(isDrawContinue) { "카드를 받을 수 없습니다" }
-        userCards.addCard(card)
-    }
-
-    fun cardSize(): Int {
-        return userCards.size
-    }
-
-    fun calculateCardPoints(): Int {
-        return userCards.calculatePoints()
-    }
-
-    fun stopCardDraw() {
-        isDrawContinue = false
-    }
-
-    fun continueCardDraw() {
-        isDrawContinue = true
-    }
-
-    fun findAllCardsNames(): List<String> {
-        return userCards.map { card ->
-            card.rank.keyword + card.suits.koreanName
-        }
-    }
-}
-
-data class Players(val players: List<Player>) : Collection<Player> by players {
-    fun getPlayerNames(): List<String> {
-        return players.map { it.name }
-    }
-
-    companion object {
-        private const val DELIMITER = ","
-
-        fun create(playerNames: String): Players {
-            val players = playerNames.split(DELIMITER).map { playerName -> Player(playerName) }
-            return Players(players)
-        }
-    }
-}
-
-fun main() {
-    val players = InputView.enterParticipatingPlayers()
-    val gameCards = GameCards.create()
-    val blackJackGame = BlackJackGame(players, gameCards)
-    blackJackGame.startGame()
-    OutputView.printFirstAllPlayersCards(players)
-    players.forEach { player ->
-        blackJackGame.handlePlayerDraw(player, gameCards)
-    }
-    OutputView.printFinalResults(players)
 }
